+++ CHANGELOG.md
@@ -13,10 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 2019-10-10
 ### Changed
-- code cleanup and improvements in controller  and view in spring-gui-jdbc
+- code cleanup and improvements in controller and view for spring-gui-jdbc and spring-gui-jpa
 
 ### Added
-- new tests for view in spring-gui-jdbc
+- new tests for view in spring-gui-jdbcc and spring-gui-jpa
 
 
 ## 2019-10-07